foldering: CoreNetwork 폴더 구조 변경 [IOS-23]
