+++ Projects/Core/Network/Sources/API/Auth/Service/AuthAPIService.swift
@@ -10,7 +10,13 @@ import Moya
 
 public final class AuthAPIService: APIServiceProtocol {
   public typealias Router = AuthAPIRouter
-  public let provider: Moya.MoyaProvider<AuthAPIRouter> = .init()
+  public let provider: Moya.MoyaProvider<AuthAPIRouter> = .init(plugins: [
+    MoyaLoggingPlugin(),
+    AccessTokenPlugin { target in
+      let accesstoken = "accesstoken"
+      return accesstoken
+    }
+  ])
   
   public init() { }
 }
