+++ Projects/Core/Network/Sources/DTO/Auth/Mobile/MobileResponseDTO.swift
