+++ Projects/Core/Network/Sources/DTO/Auth/Token/TokenRequestDTO.swift
@@ -1,0 +1,12 @@
+//
+//  TokenRequestDTO.swift
+//  CoreNetwork
+//
+//  Created by 윤지호 on 1/17/24.
+//
+
+import Foundation
+
+public struct TokenRequestDTO: Encodable {
+ let accessToken: String
+}
