--- Projects/Core/Network/Sources/Services/User/DTO/UserUpdate/UserMBTIRequestDTO.swift
@@ -1,8 +1,0 @@
-//
-//  UserMBTIRequestDTO.swift
-//  CoreNetworkInterface
-//
-//  Created by 윤지호 on 1/15/24.
-//
-
-import Foundation
