--- Projects/Core/Network/Sources/Services/User/DTO/UserUpdate/UserUpdateRequestDTO.swift
@@ -1,21 +1,0 @@
-//
-//  UserUpdateRequestDTO.swift
-//  CoreNetworkInterface
-//
-//  Created by 윤지호 on 1/15/24.
-//
-
-import Foundation
-
-public struct UserUpdateRequestDTO: Encodable {
-  let coordinate: Coordinate
-  let nickname: String
-  let gender: String
-  let birth: String
-  let mbti: String
-  
-  struct Coordinate: Encodable {
-    let lat: Double
-    let lng: Double
-  }
-}
